+++ tests/dataframe-test.js
@@ -544,11 +544,11 @@ test('DataFrame rows can be ', (assert) => {
 
     assert.deepEqual(
         df3.sortBy('id', true).toArray(), [
-            [8, 4],
             [8, 1],
+            [8, 4],
             [6, 0],
-            [3, 2],
             [3, 1],
+            [3, 2],
             [1, 1],
             [1, 1],
         ], 'sorted and reverse by a column.'
@@ -561,11 +561,11 @@ test('DataFrame rows can be ', (assert) => {
 
     assert.deepEqual(
         df3.union(df5).toArray(), [
-            [8, 4],
             [8, 1],
+            [8, 4],
             [6, 0],
-            [3, 2],
             [3, 1],
+            [3, 2],
             [1, 1],
             [1, 1],
             [3, 1],